+++ src/openfermion/hamiltonians/_hubbard.py
@@ -15,8 +15,8 @@ from __future__ import absolute_import
 
 from openfermion.ops import (FermionOperator,
                              hermitian_conjugated,
-                             number_operator,
-                             up_index, down_index)
+                             number_operator)
+from openfermion.hamiltonians import up_index, down_index
 
 
 def fermi_hubbard(x_dimension, y_dimension, tunneling, coulomb,
